Breaking settings/link view and the viewclass Controller into a component
