+++ ember-app/app/controllers/settings/links.js
@@ -1,10 +1,9 @@
 import Ember from 'ember';
 
-var SettingsLinksController = Ember.ArrayController.extend({
+var SettingsLinksController = Ember.Controller.extend({
   needs: ["settings"],
   repository: Ember.computed.alias("controllers.settings.model.repository"),
   columns: Ember.computed.alias("controllers.settings.model.column_labels"),
-  itemController: 'settings.link',
   shouldDisplayWarning: function(){
     return this.get("content.length") > 5;
   }.property('content.length')
